requirements.txt
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
astroid==2.11.7
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.6.0
black==22.6.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Deprecated==1.2.13
dill==0.3.5.1
flake8==4.0.1
frozenlist==1.3.1
idna==3.3
imageio-ffmpeg==0.4.7
isort==5.10.1
lazy-object-proxy==1.7.1
lxml==4.9.1
mccabe==0.6.1
minimap-renderer==0.0.3
multidict==6.0.2
mypy-extensions==0.4.3
nextcord==2.1.0
numpy==1.23.1
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
polib==1.1.1
pycodestyle==2.8.0
pycryptodomex==3.15.0
pyflakes==2.4.0
pylint==2.14.5
pyparsing==3.0.9
python-dotenv==0.20.0
redis==4.3.4
rq==1.11.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
tqdm==4.64.0
typing_extensions==4.3.0
wrapt==1.14.1
yarl==1.8.1
